Tagging v0.21

+++ mplayer/libavformat/cutils.c
@@ -19,13 +19,12 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 #include "avformat.h"
-#include "internal.h"
 
 /* add one element to a dynamic array */
-void ff_dynarray_add(intptr_t **tab_ptr, int *nb_ptr, intptr_t elem)
+void ff_dynarray_add(unsigned long **tab_ptr, int *nb_ptr, unsigned long elem)
 {
     int nb, nb_alloc;
-    intptr_t *tab;
+    unsigned long *tab;
 
     nb = *nb_ptr;
     tab = *tab_ptr;
@@ -34,7 +33,7 @@ void ff_dynarray_add(intptr_t **tab_ptr, int *nb_ptr, intptr_t elem)
             nb_alloc = 1;
         else
             nb_alloc = nb * 2;
-        tab = av_realloc(tab, nb_alloc * sizeof(intptr_t));
+        tab = av_realloc(tab, nb_alloc * sizeof(unsigned long));
         *tab_ptr = tab;
     }
     tab[nb++] = elem;
